Update initialization scripts

+++ .scripts/init.sh
@@ -6,13 +6,38 @@ if [[ $PWD != $HOME ]]; then
     exit 1
 fi
 
-# Perform OS-specific initialization.
+shopt -s expand_aliases
+
+source ~/.shell_aliases
 source ~/.shell_functions
+
+# Perform OS-specific initialization.
 source ~/.scripts/init-$(lowercase `uname`).sh
 
 # Install system-wide Python packages.
-if [ -f ~/requirements.txt ]; then
-    pip3 install -r ~/requirements.txt
+if exists pip3; then
+    echo "[$(sdate)] – Starting Python init."
+    pip3 install --upgrade pip 
+
+    if [ -f ~/requirements.txt ]; then
+        pip3 install -r ~/requirements.txt
+    fi
+
+    echo "[$(sdate)] – Finished Python init."
+else 
+    echo "[$(sdate)] – Skipping Python init."
+fi
+
+# Install Rust, including clippy and rustfmt.
+if ! exists rustup; then
+    echo "[$(sdate)] – Starting Rust init."
+    curl --proto '=https' --tlsv1.2 -sSf https://sh.rustup.rs | bash /dev/stdin --no-modify-path
+
+    rustup component add clippy
+    rustup component add rustfmt
+    echo "[$(sdate)] – Finished Rust init."
+else 
+    echo "[$(sdate)] – Skipping Rust init."
 fi
 
 # Remove unnecessary files (locally, not remotely).
